backend/requirements.txt
fastapi==0.116.1
uvicorn==0.18.3
sqlalchemy==2.0.20
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator>=2.0.0
pydantic==2.11.7
typing-inspection==0.4.1
colorama==0.4.6
click==8.2.1
anyio==4.9.0
h11==0.16.0
sniffio==1.3.1
starlette>=0.40.0,<0.48.0
python-multipart==0.0.20
PyJWT==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0
sendgrid==6.10.0
requests==2.31.0
psycopg2-binary==2.9.9 